vim: remove caveats
I backported a patch from ruby-trunk that fixes the shared library
reference to _environ, which is disallowed on OS X in favor of
_NSGetEnviron(). This allows vim to work with Ruby 1.9 (from Homebrew),
at least in my cursory testing.

So if you have a Homebrew-built ruby, reinstall it and vim should
compile against it correctly. You'll have to recompile e.g. Command-T
against the newer ruby as well.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>